requirements.txt
web3==5.28.0
pysha3==1.0.2
bitcoin==1.1.42